+++ Paco/res/layout/experiment_detail_part.xml
@@ -42,9 +42,14 @@
 			<LinearLayout android:layout_width="wrap_content"
                 android:layout_height="wrap_content" android:orientation="vertical"
                 android:layout_gravity="left">
-                <LinearLayout android:layout_width="wrap_content"
-                    android:layout_height="wrap_content" android:orientation="vertical"
-                    android:layout_gravity="left">
+
+                <LinearLayout
+                    android:id="@+id/scheduleDisplayPanel"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="left"
+                    android:orientation="vertical" >
+
                     <TextView android:layout_width="wrap_content"
                         android:layout_height="wrap_content" android:textStyle="bold"
                         android:text="@string/schedule" android:layout_gravity="left"></TextView>
