+++ Shared/src/com/google/paco/shared/model/ExperimentDAOCore.java
@@ -4,8 +4,6 @@ import java.io.Serializable;
 
 public class ExperimentDAOCore implements Serializable {
 
-  public static final int SCHEDULED_SIGNALING = 1;
-  public static final int TRIGGERED_SIGNALING = 1;
   protected String title;
   protected String description;
   protected String informedConsentForm;
@@ -15,11 +13,9 @@ public class ExperimentDAOCore implements Serializable {
   protected String endDate;
   protected String joinDate;
   protected Long id;
-  protected SignalingMechanismDAO[] signalingMechanisms;
-  protected SignalScheduleDAO schedule;
 
   public ExperimentDAOCore(Long id, String title, String description, String informedConsentForm,
-                           String email, SignalingMechanismDAO[] signalingMechanisms, Boolean fixedDuration, 
+                           String email, Boolean fixedDuration, 
                            String startDate, String endDate, String joinDate) {
     super();
     this.id = id;
@@ -27,8 +23,6 @@ public class ExperimentDAOCore implements Serializable {
     this.description = description;
     this.informedConsentForm = informedConsentForm;
     this.creator = email;
-    this.signalingMechanisms = signalingMechanisms;
-    setScheduleForBackwardCompatibility();
     this.fixedDuration = fixedDuration;
     this.startDate = startDate;
     this.endDate = endDate;
@@ -40,8 +34,6 @@ public class ExperimentDAOCore implements Serializable {
    */
   public ExperimentDAOCore() {
     super();
-    this.signalingMechanisms = new SignalingMechanismDAO[] { new SignalScheduleDAO()};
-    setScheduleForBackwardCompatibility();
   }
 
   public String getTitle() {
@@ -114,33 +106,6 @@ public class ExperimentDAOCore implements Serializable {
 
   public void setId(Long id) {
     this.id = id;
-  }
-
-  public SignalingMechanismDAO[] getSignalingMechanisms() {
-    return signalingMechanisms;
-  }
-
-  public void setSignalingMechanisms(SignalingMechanismDAO[] signalingMechanisms) {
-    this.signalingMechanisms = signalingMechanisms;
-  }
-
-  public void setScheduleForBackwardCompatibility() {
-    if (getSignalingMechanisms() != null 
-            && getSignalingMechanisms().length > 0 
-            && getSignalingMechanisms()[0] instanceof SignalScheduleDAO) {
-      schedule = (SignalScheduleDAO) getSignalingMechanisms()[0];
-    } else {
-      schedule = new SignalScheduleDAO();
-      schedule.setScheduleType(SignalScheduleDAO.SELF_REPORT);
-    }
-  }
-
-  public SignalScheduleDAO getSchedule() {
-    return schedule;
-  }
-
-  public void setSchedule(SignalScheduleDAO schedule) {
-    this.schedule = schedule;
   }
 
 }
